Simple implementation of concurrent write for independent partitions.
The implementation have following limitations:
 * if transaction reads from a set of partitions it must write to the
   same partition set (If TX reads from P1, P2 then it must write to P1, P2)
   This condition is satisfied for update, replaceWhere, merge
   (if partitioning is not changed in data)
 * transactions can't perform delete operation

Delete operation is not supported, because RemoveFile doesn't hold info
about partitions. Technically, it would be possible to extract this
information from file path, but its not a nice solution.

Since OptimisticTransaction doesn't track which transaction was read
we need to have constraint for (partitions read) == (partitions written).
In this way its possible to check conflicts only for written partitions,
since read ones are the same set.